Add files via upload
http://programmingbydoing.com/a/nesting-loops.html

What You Should Do on Your Own
Assignments turned in without these things will receive no credit.

Look at the first set of nested loops ("CN"). Which variable changes faster? Is it the variable controlled by the outer loop (c) or the variable controlled by the inner loop (n)? Answer in a comment.
Change the order of the loops so that the "c" loop is on the inside and the "n" loop is on the outside. How does the output change?
Look at the second set of nested loops ("AB"). Change the print() statement to println(). How does the output change? (Then change it back to print().)
Add a System.out.println() statement after the close brace of the inner loop (the "b" loop), but still inside the outer loop. How does the output change?